+++ app/Http/Controllers/MeasureController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Requests\MeasureRequest;
 use App\Models\Measure;
+use Illuminate\Support\Facades\DB;
 
 class MeasureController extends Controller
 {
@@ -47,15 +48,29 @@ class MeasureController extends Controller
         return $response;
     }
 
+
+
+
     public function destroy($id)
     {
-        $query = Measure::where('id_measure', $id)->delete(); 
 
-        if($query){
-            $response = array('status' => 1, 'msg'=>'Deleted');
+        $product = DB::table('products')
+                    ->join('measure', 'measure.id_measure', '=', 'products.id_measure')
+                    ->select('products.id_products AS idProduct') 
+                    ->where('products.id_measure', '=', $id)
+                    ->get();
+
+        if(count($product) == 0){
+            $query = Measure::where('id_measure', $id)->delete(); 
+
+            if($query){
+                $response = array('status' => 1, 'msg'=>'Deleted');
+            }else{
+                $response = array('status' => 0, 'msg'=>'Fail');
+            }
         }else{
-            $response = array('status' => 0, 'msg'=>'Fail');
-        }
+            $response = array('status' => 409, 'msg'=>'Fail, foreign key reference');
+        }      
         
         return Response()->json($response); 
     }
